backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.1
gunicorn==22.0.0
install==1.3.5
psycopg2==2.9.9
sqlparse==0.5.0